+++ src/components/app/Page.svelte
@@ -1,19 +1,34 @@
 <script lang="ts">
+    import { goto } from '$app/navigation';
+    import { page } from '$app/stores';
+    import Button from '../widgets/Button.svelte';
     import Settings from '../settings/Settings.svelte';
-    import Link from './Link.svelte';
+    import { locale } from '../../db/Database';
 
     export let fullscreen = false;
+
+    function handleKey(event: KeyboardEvent) {
+        if (
+            (event.ctrlKey || event.metaKey) &&
+            event.key === 'Escape' &&
+            $page.route.id !== null
+        ) {
+            goto('/');
+        }
+    }
 </script>
+
+<svelte:window on:keydown={handleKey} />
 
 <div class="page">
     <main>
         <slot />
     </main>
     <footer class:fullscreen>
-        <div class="beta"
-            ><Link external to="https://github.com/amyjko/wordplay/milestone/1"
-                >beta</Link
-            ></div
+        <Button
+            tip={$locale.ui.widget.back}
+            active={$page.route.id !== '/'}
+            action={() => history.back()}><span class="back">⏴</span></Button
         >
         <Settings />
     </footer>
@@ -29,10 +44,6 @@
         bottom: 0;
         display: flex;
         flex-direction: column;
-    }
-
-    .beta {
-        font-style: italic;
     }
 
     main {
@@ -61,9 +72,15 @@
             var(--wordplay-border-width);
         background: var(--wordplay-background);
         z-index: 1;
+        gap: var(--wordplay-spacing);
     }
 
     .fullscreen:not(:hover) {
         opacity: 0.25;
+    }
+
+    .back {
+        display: inline-block;
+        font-size: 150%;
     }
 </style>
