+++ src/locale/en-US.json
@@ -3692,7 +3692,8 @@
             },
             "loading": {
                 "message": "Loading font faces and text, thanks for waiting!"
-            }
+            },
+            "back": "back to previous page"
         },
         "tile": {
             "toggle": {
